ci: auto update binary dependencies
This PR adds automation to support upgrading SOCI and ECR Cred helper
dependencies to their latest versions. Finch hardcodes the SHA of the
ECR Credential Helper binary for the arm64 architecture to be used in
checking for if it has been `Installed()`. However, this logic likely
needs to be reworked as Finch supports both x86_64 and ARM
architectures. Checking for previously installed ECR cred helper can be
done in a separate PR.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>